+++ hana/lib/HANAService.js
@@ -5,7 +5,6 @@ const { Readable } = require('stream')
 const { SQLService } = require('@cap-js/db-service')
 const drivers = require('./drivers')
 const cds = require('@sap/cds')
-const { getTransition } = require('@sap/cds/libx/_runtime/common/utils/resolveView')
 const collations = require('./collations.json')
 const sessionVariableMap = require('./session.json')
 const keywords = cds.compiler.to.hdi.keywords
@@ -752,28 +751,20 @@ class HANAService extends SQLService {
     INSERT_entries(q) {
       this.values = undefined
       const { INSERT } = q
+      // REVISIT: should @cds.persistence.name be considered ?
+      const entity = q._target?.['@cds.persistence.name'] || this.name(q._target?.name || INSERT.into.ref[0], q)
 
       const elements = q.elements || q._target?.elements
       if (!elements) {
         return super.INSERT_entries(q)
       }
 
-      const entity = q._target ? this.table_name(q) : INSERT.into.ref[0]
-      // REVISIT: remove fallback when cds.dbs requires cds >= 9.3
-      const transitions = this.srv.resolve.transitions4db ? this.srv.resolve.transitions4db(q) : getTransition(q._target, this.srv, false, 'INSERT')
-
       const columns = elements
-        ? ObjectKeys(elements).filter(c => (c = transitions.mapping.get(c)?.ref?.[0] || c)
-          && c in transitions.target.elements
-          && !transitions.target.elements[c].virtual
-          && !transitions.target.elements[c].value
-          && !transitions.target.elements[c].isAssociation
-          && !elements[c]?.[SYSTEM_VERSIONED]
-        )
+        ? ObjectKeys(elements).filter(c => c in elements && !elements[c].virtual && !elements[c].value && !elements[c].isAssociation && !elements[c][SYSTEM_VERSIONED])
         : ObjectKeys(INSERT.entries[0])
       this.columns = columns
 
-      const extractions = this._managed = this.managed(columns.map(c => ({ name: c })), elements).slice(0, columns.length)
+      const extractions = this.managed(columns.map(c => ({ name: c })), elements)
 
       // REVISIT: @cds.extension required
       const extraction = extractions.map(c => c.extract)
@@ -813,8 +804,9 @@ class HANAService extends SQLService {
       // With the buffer table approach the data is processed in chunks of a configurable size
       // Which allows even smaller HANA systems to process large datasets
       // But the chunk size determines the maximum size of a single row
-      return (this.sql = `INSERT INTO ${this.quote(entity)} (${this.columns.map(c => this.quote(transitions.mapping.get(c)?.ref?.[0] || c))
-        }) WITH SRC AS (SELECT ? AS JSON FROM DUMMY UNION ALL SELECT TO_NCLOB(NULL) AS JSON FROM DUMMY)
+      return (this.sql = `INSERT INTO ${this.quote(entity)} (${this.columns.map(c =>
+        this.quote(c),
+      )}) WITH SRC AS (SELECT ? AS JSON FROM DUMMY UNION ALL SELECT TO_NCLOB(NULL) AS JSON FROM DUMMY)
       SELECT ${converter} FROM JSON_TABLE(SRC.JSON, '$' COLUMNS(${extraction}) ERROR ON ERROR) AS NEW`)
     }
 
@@ -1179,7 +1171,7 @@ SELECT ${mixing} FROM JSON_TABLE(SRC.JSON, '$' COLUMNS(${extraction}) ERROR ON E
     }
 
     managed_extract(name, element, converter) {
-      const path = this.string(this.srv.server.major <= 2 ? `$.${name}` : `$[${JSON.stringify(name)}]`)
+      const path = this.string(this.srv?.server.major <= 2 ? `$.${name}` : `$[${JSON.stringify(name)}]`)
       return {
         extract: `${this.quote(name)} ${this.insertType4(element)} PATH ${path}, ${this.quote('$.' + name)} NVARCHAR(2147483647) FORMAT JSON PATH ${path}`,
         sql: converter(`NEW.${this.quote(name)}`),